added wishlist session integration

+++ src/main/java/org/example/oenskelisten/Controller/WishListController.java
@@ -1,5 +1,7 @@
 package org.example.oenskelisten.Controller;
 
+import jakarta.servlet.http.HttpSession;
+import org.example.oenskelisten.Model.User;
 import org.example.oenskelisten.Model.Wish;
 import org.example.oenskelisten.Model.WishList;
 import org.example.oenskelisten.Service.WishListService;
@@ -37,10 +39,15 @@ public class WishListController {
 
 
     @GetMapping("/list/{id}")
-    public String getWishList(@PathVariable("id") int id, Model model) {
+    public String getWishList(@PathVariable("id") int id, Model model, HttpSession session) {
         System.out.println("wishlist list id: " + id);
         model.addAttribute("pageID", id);
+
         WishList wishList = wishListService.getWishListModelByID(id);
+
+        //checking to see if we are the owner of the wishlist:
+        boolean userIsOwner = isUserOwner(session, wishList.getUserID());
+        model.addAttribute("userIsOwner", userIsOwner);
         model.addAttribute(wishList);
         return "grid-wishlist";
     }
@@ -60,9 +67,14 @@ public class WishListController {
     }
 
     @PostMapping("/save")
-    public String addWish(@ModelAttribute("wish") Wish wish) {
+    public String addWish(@ModelAttribute("wish") Wish wish, HttpSession session) {
         String destination = "wishlist/list/" + wish.getWishlistID();
-        wishListService.addWish(wish);
+
+        WishList wishList = wishListService.getWishListModelByID(wish.getWishlistID());
+        if(isUserOwner(session, wishList.getUserID())){
+            wishListService.addWish(wish);
+        }
+
         return "redirect:/" + destination;
     }
 
@@ -74,8 +86,11 @@ public class WishListController {
     }
 
     @PostMapping("/update")
-    public String editWish(@ModelAttribute("wish") Wish wish) {
-        wishListService.updateWish(wish);
+    public String editWish(@ModelAttribute("wish") Wish wish, HttpSession session) {
+        WishList wishList = wishListService.getWishListModelByID(wish.getWishlistID());
+        if(isUserOwner(session, wishList.getUserID())){
+            wishListService.updateWish(wish);
+        }
 
         return "redirect:/" + "wishlist/list/" + wish.getWishlistID();
     }
@@ -83,13 +98,26 @@ public class WishListController {
     //the only things sent to deletion correctly are wish id and wishlist id. if you need more
     //you have to change the grid-wishlist html code
     @PostMapping("/delete")
-    public String deleteWish(@ModelAttribute("wish") Wish formWish) {
-        System.out.println(formWish.getId());
+    public String deleteWish(@ModelAttribute("wish") Wish formWish, HttpSession session) {
         Wish wish = wishListService.getWishById(formWish.getId());
-        if (wish == null) {
-            throw new IllegalArgumentException("Wish does not exist");
+        WishList wishList = wishListService.getWishListModelByID(wish.getWishlistID());
+        if(isUserOwner(session, wishList.getUserID())){
+            if (wish == null) {
+                throw new IllegalArgumentException("Wish does not exist");
+            }
+            wishListService.deleteWishById(wish.getId());
         }
-        wishListService.deleteWishById(wish.getId());
+
         return "redirect:/wishlist/list/" + formWish.getWishlistID();
+    }
+
+    private boolean isUserOwner(HttpSession session, int ownerID){
+        var hopefullyAUSer = session.getAttribute("user");
+        if(hopefullyAUSer instanceof User loggedInUser){
+            if( loggedInUser.getUserID() == ownerID){
+                return true;
+            }
+        }
+        return false;
     }
 }
